refactor: move `PosterCard` to `ui/`
