+++ frontend/components/ui/poster-card.tsx
